requirements-test.txt
# 测试依赖
pytest>=7.0.0
pytest-qt>=4.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xvfb>=3.0.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0

# 测试工具
coverage>=7.2.0
pytest-html>=3.1.1
pytest-benchmark>=4.0.0
pytest-profiling>=1.7.0

# 模拟和桩
responses>=0.23.0
faker>=18.0.0
factory-boy>=3.2.0

# 性能测试
memory-profiler>=0.60.0
line-profiler>=4.0.0
py-spy>=0.3.14

# 质量检查
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.2.0

# UI 测试
PyQt6-stubs>=6.5.0.3.2
qtawesome>=1.2.3